requirements.txt
streamlit==1.5.0
pyperclip==1.8.2
qrcode==7.3.1
Pillow==8.4.0
pyzbar==0.1.9
opencv-python==4.5.5.62